add an option to disable cache control headers
While the options st offers for cache control are good
enough for most use-cases, there are some that it is not
able to handle. And since st ends the request after serving
a file, the user has no way to customize cache headers
any further.

This commit makes it such that if you set cache.content.maxAge
to `false`, it does not set cache control headers, allowing
the user to set them beforehand without the headers being
overwritten by st.