Report actual address we're listening on
Telling the user that we're listening (only) on 127.0.0.1 when in fact we
are listening on all available addresses can confuse users into thinking
that running the service is more secure than it actually is.  Better to tell
the user what host exactly we are listening on.  As far as I can tell,
browsers accept having the zero address (:: or 0.0.0.0) pasted into their
location bar, so it should still be possible to copy & paste the URL.

One benefit is that we now only report the listening line when we have
actually started listening, avoiding confusion when the web server is slow
to start or when it fails to start after the message got printed.  Another
benefit is that we can report the port number actually in use if the user
set the port to zero.