fix: change `nvm install current` to `nvm install node`
`nvm install current` is deprectaed and instead we need to use `nvm install node`

The Travis CI is failing because of that. Any new pull requests in this project will fail
fail in Travis pipeline because of this change.

An example could be found here #93

I am keeping this PR separately to get a track why we changed this `yml` file.