Improve regular expression for removing trailing slashes
Update the regular expression that is used to remove trailing slashes
from the path in `Mount#getPath` to avoid super-linear runtimes. I
believe this does not affect the `st` export because the only path to
this API is through `Mount#serve` which sanitizes the `req.url` using
`path.normalize` (in `Mount#getUriPath`). However, the `Mount` class
is also exported directly, through which `getPath` could be called on
unsanitized inputs.

The super-linear runtime could occur because the regex engine will
backtrack through a repetition of '/'s if it is not at the end. This
is resolved by anchoring a repetition '/' on a non-'/' with a negative
lookbehind.

The issue can be tested and reproduced using this script:

    const st = require('st')
    const path = require('path')
    const mount = new st.Mount({
      path: path.join(__dirname, '/static'),
      url: '/static'
    })

    for (let i = 0; i < 90_000; i += 10_000) {
      console.time('timer')
      mount.getPath(`a${"/".repeat(i)}a`)
      console.timeEnd('timer')
    }